+++ src/east/__main__.py
@@ -3,7 +3,7 @@ import rich_click
 
 from .east_context import EastContext, east_group_settings
 from .system_commands import sys_setup, util
-from .workspace_commands import build, bypass, clean, flash, release, update
+from .workspace_commands import build, bypass, clean, debug, flash, release, update
 
 rich_click.rich_click.MAX_WIDTH = 80
 rich_click.rich_click.USE_RICH_MARKUP = True
@@ -13,7 +13,15 @@ rich_click.rich_click.COMMAND_GROUPS = {
     "east": [
         {
             "name": "Workspace commands",
-            "commands": ["build", "flash", "clean", "update", "bypass", "release"],
+            "commands": [
+                "build",
+                "flash",
+                "clean",
+                "debug",
+                "update",
+                "bypass",
+                "release",
+            ],
         },
         {
             "name": "System commands",
@@ -68,6 +76,7 @@ cli.add_command(update)
 cli.add_command(sys_setup)
 cli.add_command(util)
 cli.add_command(release)
+cli.add_command(debug)
 
 
 def main():